+++ internal/builder/controllerbuilder/controller_app_test.go
@@ -105,7 +105,7 @@ func TestBuilder_BuildController(t *testing.T) {
 			require.Equal(t, common.SlurmUserGid, ptr.Deref(got.Spec.Template.Spec.Containers[0].SecurityContext.RunAsGroup, 0))
 			require.Equal(t, labels.ControllerApp, got.Spec.Template.Spec.Containers[0].Name)
 			require.Equal(t, labels.ControllerApp, got.Spec.Template.Spec.Containers[0].Ports[0].Name)
-			require.Equal(t, common.SlurmctldPort, got.Spec.Template.Spec.Containers[0].Ports[0].ContainerPort)
+			require.Equal(t, int32(common.SlurmctldPort), got.Spec.Template.Spec.Containers[0].Ports[0].ContainerPort)
 		})
 	}
 }
